Seperating config into a1_env_cfg file.
